moved document related code to UI file
