+++ .eslintrc.js
@@ -18,5 +18,6 @@ module.exports = {
     'arrow-parens': 0,
     'no-param-reassign': 0,
     'no-use-before-define': 0,
+    'import/no-cycle': 0,
   },
 };
