+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "eslint": "eslint --fix"
+    "eslint": "eslint --fix",
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "live-server"
   },
