+++ yarn-error.log
@@ -1,0 +1,59 @@
+Arguments: 
+  /usr/bin/node /usr/share/yarn/bin/yarn.js start
+
+PATH: 
+  /home/shamim/.rbenv/plugins/ruby-build/bin:/home/shamim/.rbenv/shims:/home/shamim/.rbenv/bin:/home/shamim/.rbenv/plugins/ruby-build/bin:/home/shamim/.rbenv/shims:/home/shamim/.rbenv/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games:/usr/local/games:/snap/bin
+
+Yarn version: 
+  1.17.3
+
+Node version: 
+  12.10.0
+
+Platform: 
+  linux x64
+
+Trace: 
+  SyntaxError: /home/shamim/Microverse/Microverse/JavaScript/Project/Tic-Tac-Toe/package.json: Unexpected string in JSON at position 140
+      at JSON.parse (<anonymous>)
+      at /usr/share/yarn/lib/cli.js:1625:59
+      at Generator.next (<anonymous>)
+      at step (/usr/share/yarn/lib/cli.js:304:30)
+      at /usr/share/yarn/lib/cli.js:315:13
+
+npm manifest: 
+  {
+    "name": "tic-tac-toe",
+    "version": "1.0.0",
+    "description": "",
+    "main": "index.js",
+    "scripts": {
+      "eslint": "eslint --fix"
+      "test": "echo \"Error: no test specified\" && exit 1",
+      "start": "live-server"
+    },
+    "repository": {
+      "type": "git",
+      "url": "git+https://github.com/shshamim63/Tic-Tac-Toe.git"
+    },
+    "author": "",
+    "license": "ISC",
+    "bugs": {
+      "url": "https://github.com/shshamim63/Tic-Tac-Toe/issues"
+    },
+    "homepage": "https://github.com/shshamim63/Tic-Tac-Toe#readme",
+    "devDependencies": {
+      "eslint": "^6.3.0",
+      "eslint-config-airbnb-base": "^14.0.0",
+      "eslint-plugin-import": "^2.18.2"
+    },
+    "dependencies": {
+      "live-server": "^1.2.1"
+    }
+  }
+
+yarn manifest: 
+  No manifest
+
+Lockfile: 
+  No lockfile
